extracted buisness login form utils
- moved all the business logic from utils
- fixed bug in e2e tests
- divided unit tests to packages
