+++ internals/finalizer/finalizer.go
@@ -11,10 +11,11 @@ import (
 
 const FinalizerCleanupCapp = "dana.io/capp-cleanup"
 
-func HandleResourceDeletion(ctx context.Context, capp rcsv1alpha1.Capp, r client.Client, resource_managers []rmanagers.ResourceManager) (error, bool) {
+// HandleResourceDeletion manages the Capp deletion.
+func HandleResourceDeletion(ctx context.Context, capp rcsv1alpha1.Capp, r client.Client, resourceManagers map[string]rmanagers.ResourceManager) (error, bool) {
 	if capp.ObjectMeta.DeletionTimestamp != nil {
 		if controllerutil.ContainsFinalizer(&capp, FinalizerCleanupCapp) {
-			if err := finalizeService(capp, resource_managers); err != nil {
+			if err := finalizeService(capp, resourceManagers); err != nil {
 				return err, false
 			}
 			return RemoveFinalizer(ctx, capp, r), true
@@ -23,6 +24,7 @@ func HandleResourceDeletion(ctx context.Context, capp rcsv1alpha1.Capp, r client
 	return nil, false
 }
 
+// RemoveFinalizer removes the finalizer from the Capp manifest.
 func RemoveFinalizer(ctx context.Context, capp rcsv1alpha1.Capp, r client.Client) error {
 	controllerutil.RemoveFinalizer(&capp, FinalizerCleanupCapp)
 	if err := r.Update(ctx, &capp); err != nil {
@@ -31,8 +33,9 @@ func RemoveFinalizer(ctx context.Context, capp rcsv1alpha1.Capp, r client.Client
 	return nil
 }
 
-func finalizeService(capp rcsv1alpha1.Capp, resource_managers []rmanagers.ResourceManager) error {
-	for _, manager := range resource_managers {
+// fnializeService runs the cleanup of all the resource mangers.
+func finalizeService(capp rcsv1alpha1.Capp, resourceManagers map[string]rmanagers.ResourceManager) error {
+	for _, manager := range resourceManagers {
 		if err := manager.CleanUp(capp); err != nil {
 			return err
 		}
@@ -40,7 +43,7 @@ func finalizeService(capp rcsv1alpha1.Capp, resource_managers []rmanagers.Resour
 	return nil
 }
 
-// ensureFinalizer ensures the service has the finalizer
+// EnsureFinalizer ensures the service has the finalizer.
 func EnsureFinalizer(ctx context.Context, service rcsv1alpha1.Capp, r client.Client) error {
 	if !controllerutil.ContainsFinalizer(&service, FinalizerCleanupCapp) {
 		controllerutil.AddFinalizer(&service, FinalizerCleanupCapp)
