+++ internals/resource-managers/flow.go
@@ -3,6 +3,8 @@ package resourceprepares
 import (
 	"context"
 	"fmt"
+	"reflect"
+
 	rcsv1alpha1 "github.com/dana-team/container-app-operator/api/v1alpha1"
 	rclient "github.com/dana-team/container-app-operator/internals/wrappers"
 	"github.com/go-logr/logr"
@@ -13,7 +15,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
-	"reflect"
+
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -66,21 +68,20 @@ func (f FlowManager) prepareResource(capp rcsv1alpha1.Capp) loggingv1beta1.Flow 
 // CleanUp deletes the flow resource associated with the Capp object.
 // The flow resource is deleted by calling the DeleteResource method of the resourceManager object.
 func (f FlowManager) CleanUp(capp rcsv1alpha1.Capp) error {
-
-	if f.isRequired(capp) {
+	if f.IsRequired(capp) {
 		flowName := capp.GetName() + "-flow"
 		resourceManager := rclient.ResourceBaseManagerClient{Ctx: f.Ctx, K8sclient: f.K8sclient, Log: f.Log}
 		flow := loggingv1beta1.Flow{}
 		if err := resourceManager.DeleteResource(&flow, flowName, capp.Namespace); err != nil {
-			return fmt.Errorf("unable to delete flow %s: %s", flowName, err.Error())
+			return fmt.Errorf("unable to delete flow %q: %w", flowName, err)
 		}
 
 	}
 	return nil
 }
 
-// isRequired responsible to determine if resource logging operator flow is required.
-func (f FlowManager) isRequired(capp rcsv1alpha1.Capp) bool {
+// IsRequired is responsible to determine if resource logging operator flow is required.
+func (f FlowManager) IsRequired(capp rcsv1alpha1.Capp) bool {
 	if capp.Spec.LogSpec != (rcsv1alpha1.LogSpec{}) {
 		return capp.Spec.LogSpec.Type == LogTypeElastic || capp.Spec.LogSpec.Type == LogTypeSplunk
 	}
@@ -93,7 +94,7 @@ func (f FlowManager) CreateOrUpdateObject(capp rcsv1alpha1.Capp) error {
 	flowName := capp.GetName() + "-flow"
 	logger := f.Log.WithValues("FlowName", flowName, "FlowNamespace", capp.Namespace)
 
-	if f.isRequired(capp) {
+	if f.IsRequired(capp) {
 		generatedFlow := f.prepareResource(capp)
 		// get instance of current flow
 		currentFlow := loggingv1beta1.Flow{}
@@ -105,18 +106,18 @@ func (f FlowManager) CreateOrUpdateObject(capp rcsv1alpha1.Capp) error {
 
 			if err := resourceManager.CreateResource(&generatedFlow); err != nil {
 				f.EventRecorder.Event(&capp, corev1.EventTypeWarning, eventCappFlowCreationFailed, fmt.Sprintf("Failed to create flow %s for Capp %s", flowName, capp.Name))
-				return fmt.Errorf("failed to create flow %s: %s", flowName, err.Error())
+				return fmt.Errorf("failed to create flow %q: %w", flowName, err)
 			}
 			logger.Info("Created flow successfully")
 			f.EventRecorder.Event(&capp, corev1.EventTypeNormal, eventCappFlowCreated, fmt.Sprintf("Created flow %s for Capp %s", flowName, capp.Name))
 		case err != nil:
-			return fmt.Errorf("failed to fetch existing flow %s: %s", flowName, err.Error())
+			return fmt.Errorf("failed to fetch existing flow %q: %w", flowName, err)
 		}
 		if !reflect.DeepEqual(currentFlow.Spec, generatedFlow.Spec) {
 			currentFlow.Spec = generatedFlow.Spec
 			logger.Info("Trying to update the current flow")
 			if err := resourceManager.UpdateResource(&currentFlow); err != nil {
-				return fmt.Errorf("failed to update the current flow %s: %s", flowName, err.Error())
+				return fmt.Errorf("failed to update the current flow %q: %w", flowName, err)
 			}
 			logger.Info("Current flow successfully updated")
 		}
