+++ internals/status/application_links.go
@@ -1,4 +1,4 @@
-package status_utils
+package status
 
 import (
 	"strings"
@@ -18,9 +18,9 @@ const (
 
 // This function builds the ApplicationLinks status of the Capp  by getting the console route.
 // It returns a pointer to the ApplicationLinks struct.
-func buildApplicationLinks(ctx context.Context, log logr.Logger, r client.Client, onOpenshift bool) (*rcsv1alpha1.ApplicationLinks, error) {
+func buildApplicationLinks(ctx context.Context, log logr.Logger, r client.Client, isRequired bool) (*rcsv1alpha1.ApplicationLinks, error) {
 	applicationLinks := rcsv1alpha1.ApplicationLinks{}
-	if onOpenshift {
+	if isRequired {
 		console, err := getClusterConsole(ctx, log, r)
 		if err != nil {
 			return nil, err
