+++ internals/status/capp.go
@@ -1,21 +1,15 @@
-// Package status_utils This is a Go package that contains functions for synchronizing the status
-// of a custom resource definition (CRD)
-// called Capp with the status of the Knative service and revisions associated with it.
-// The SyncStatus function is the main function that orchestrates the synchronization process.
-package status_utils
+package status
 
 import (
 	"context"
+
 	rcsv1alpha1 "github.com/dana-team/container-app-operator/api/v1alpha1"
+	rmanagers "github.com/dana-team/container-app-operator/internals/resource-managers"
 	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-const (
-	cappEnabledState = "enabled"
 )
 
 // CreateStateStatus changes the state status by identifying changes in the manifest
@@ -28,7 +22,7 @@ func CreateStateStatus(stateStatus *rcsv1alpha1.StateStatus, cappStateFromSpec s
 
 // SyncStatus is the main function that synchronizes the status of the Capp CRD with the Knative service and revisions associated with it.
 // It gets the Capp CRD, builds the ApplicationLinks and RevisionInfo statuses, and updates the status of the Capp CRD if it has changed.
-func SyncStatus(ctx context.Context, capp rcsv1alpha1.Capp, log logr.Logger, r client.Client, onOpenshift bool) error {
+func SyncStatus(ctx context.Context, capp rcsv1alpha1.Capp, log logr.Logger, r client.Client, onOpenshift bool, resourceManagers map[string]rmanagers.ResourceManager) error {
 	cappObject := rcsv1alpha1.Capp{}
 	if err := r.Get(ctx, types.NamespacedName{Namespace: capp.Namespace, Name: capp.Name}, &cappObject); err != nil {
 		return err
@@ -39,7 +33,8 @@ func SyncStatus(ctx context.Context, capp rcsv1alpha1.Capp, log logr.Logger, r c
 		return err
 	}
 
-	knativeObjectStatus, revisionInfo, err := buildKnativeStatus(ctx, r, capp)
+	knativeServiceManger := resourceManagers[rmanagers.KnativeServing]
+	knativeObjectStatus, revisionInfo, err := buildKnativeStatus(ctx, r, capp, knativeServiceManger.IsRequired(capp))
 	if err != nil {
 		return err
 	}
@@ -47,13 +42,15 @@ func SyncStatus(ctx context.Context, capp rcsv1alpha1.Capp, log logr.Logger, r c
 	cappObject.Status.KnativeObjectStatus = knativeObjectStatus
 	cappObject.Status.RevisionInfo = revisionInfo
 
-	loggingStatus, err := buildLoggingStatus(ctx, capp, log, r)
+	FlowManager := resourceManagers[rmanagers.Flow]
+	loggingStatus, err := buildLoggingStatus(ctx, capp, log, r, FlowManager.IsRequired(capp))
 	if err != nil {
 		return err
 	}
 	cappObject.Status.LoggingStatus = loggingStatus
 
-	routeStatus, err := buildRouteStatus(ctx, r, capp)
+	DomainMappinManger := resourceManagers[rmanagers.DomainMapping]
+	routeStatus, err := buildRouteStatus(ctx, r, capp, DomainMappinManger.IsRequired(capp))
 	if err != nil {
 		return err
 	}
