+++ internals/status/route.go
@@ -1,7 +1,8 @@
-package status_utils
+package status
 
 import (
 	"context"
+
 	rcsv1alpha1 "github.com/dana-team/container-app-operator/api/v1alpha1"
 	"k8s.io/apimachinery/pkg/types"
 	knativev1beta1 "knative.dev/serving/pkg/apis/serving/v1beta1"
@@ -10,10 +11,10 @@ import (
 
 // buildRouteStatus constructs the Route Status of the Capp object in accordance to the
 // status of the corresponding DomainMapping object if such exists
-func buildRouteStatus(ctx context.Context, kubeClient client.Client, capp rcsv1alpha1.Capp) (rcsv1alpha1.RouteStatus, error) {
+func buildRouteStatus(ctx context.Context, kubeClient client.Client, capp rcsv1alpha1.Capp, isRequired bool) (rcsv1alpha1.RouteStatus, error) {
 	routeStatus := rcsv1alpha1.RouteStatus{}
 
-	if capp.Spec.RouteSpec == (rcsv1alpha1.RouteSpec{}) {
+	if !isRequired {
 		return routeStatus, nil
 	}
