--- internals/utils/secure/secure.go
@@ -1,36 +1,0 @@
-package secure
-
-import (
-	"fmt"
-	rcsv1alpha1 "github.com/dana-team/container-app-operator/api/v1alpha1"
-	rclient "github.com/dana-team/container-app-operator/internals/wrappers"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/tools/record"
-	knativev1beta1 "knative.dev/serving/pkg/apis/serving/v1beta1"
-)
-
-const (
-	eventCappSecretNotFound = "SecretNotFound"
-)
-
-// SetHttpsKnativeDomainMapping takes a Capp, Knative Domain Mapping and a ResourceBaseManager Client and sets the Knative Domain Mapping Tls based on the Capp's Https field.
-func SetHttpsKnativeDomainMapping(capp rcsv1alpha1.Capp, knativeDomainMapping *knativev1beta1.DomainMapping, resourceManager rclient.ResourceBaseManagerClient, eventRecorder record.EventRecorder) {
-	isHttps := capp.Spec.RouteSpec.TlsEnabled
-	if isHttps {
-		tlsSecret := corev1.Secret{}
-		if err := resourceManager.K8sclient.Get(resourceManager.Ctx, types.NamespacedName{Name: capp.Spec.RouteSpec.TlsSecret, Namespace: capp.Namespace}, &tlsSecret); err != nil {
-			if errors.IsNotFound(err) {
-				resourceManager.Log.Error(err, fmt.Sprintf("the tls secret %s for DomainMapping does not exist", capp.Spec.RouteSpec.TlsSecret))
-				eventRecorder.Event(&capp, corev1.EventTypeWarning, eventCappSecretNotFound, fmt.Sprintf("Secret %s for DomainMapping %s does not exist", capp.Spec.RouteSpec.TlsSecret, knativeDomainMapping.Name))
-				return
-			}
-			resourceManager.Log.Error(err, fmt.Sprintf("unable to get tls secret %s for DomainMapping", capp.Spec.RouteSpec.TlsSecret))
-		} else {
-			knativeDomainMapping.Spec.TLS = &knativev1beta1.SecretTLS{
-				SecretName: capp.Spec.RouteSpec.TlsSecret,
-			}
-		}
-	}
-}
