+++ test/k8s_tests/logger_e2e_test.go
@@ -2,6 +2,7 @@ package k8s_tests
 
 import (
 	"fmt"
+
 	mock "github.com/dana-team/container-app-operator/test/k8s_tests/mocks"
 	utilst "github.com/dana-team/container-app-operator/test/k8s_tests/utils"
 	. "github.com/onsi/ginkgo/v2"
